front-end#7 Refactor code: - separated logic and UI with a custom hook
