+++ app/client/src/@types/userLoginInfo.ts
@@ -1,0 +1,4 @@
+export type userLoginInfoType = {
+  username: string;
+  password: string;
+};
